Plot graphs relevant to transportability analysis using TADA

Plot graphs for assessment of covariate balance and results in a TADA analysis. This function currently supports mirrored histograms, SMD plots and model coefficient plots.

Usage

# S3 method for class 'transportTADA'
plot(
  x,
  type = "propensityHist",
  bins = 50,
  maxWeight = NULL,
  covariates = NULL,
  effectModifiers = NULL,
  ...
)

Arguments

x

Result from transportTADA function

type

One of "propensityHist", "propensitySMD", "participationHist", "msm". Hist produces mirrored histograms of estimated probability of treatment between treatment groups (for propensity), or of estimated participation weights between study and target data (for participation). SMD produces SMD plots of covariates between treatment groups (for propensity). msm produces plots showing confidence intervals for the model coefficients, which should have the correct standard errors.

bins

Number of bins for propensity score/participation weight histograms. This is only used for Hist.

maxWeight

Maximum participation weight value, i.g., range, that user wants to show in the participationHist.

covariates

Vector of strings indicating names of covariates in propensity model

effectModifiers

Vector of strings indicating names of effect modifiers in participation model

...

Further arguments from previous function or to pass to next function

Value

A ggplot object which contains the desired plot.