Summarize results of transportability analysis using g-computation

Returns summary object which contains the transported effect estimate and the outcome model summary object, as well as information about response and treatment variables. In the MSM summary object, the correct variance estimators are calculated.

Usage

# S3 method for class 'transportGC'
summary(object, ...)

# S3 method for class 'summary.transportGC'
print(x, out = stdout(), ...)

Arguments

object

Result from transportGC function

...

Further arguments from previous function or to pass to next function

x

summary.transportGC object.

out

Output stream.

Value

The summary.transportGC function returns a summary.transportGC object containing the following components:

msmSummary: Summary object of MSM with correct variance estimates.

preparedModelSummary: Summary object of outcome model, provided only for information. No conclusions should be drawn from the outcome model.

response: String indicating response variable name.

treatment: String indicating treatment variable name.

treatmentLevels: Vector of strings indicating levels of treatment variable