Summarize results of a fitted MSM using the TADA approach

Returns summary object which contains a summary of the fitted MSM, pre- and post-weighting standardized mean differences (SMDs) and pre-post matching table including aggregate level summary of effect modifiers of interest.

Usage

# S3 method for class 'transportTADA'
summary(object, covariates = NULL, effectModifiers = NULL, ...)

# S3 method for class 'summary.transportTADA'
print(x, out = stdout(), ...)

Arguments

object

Result from transportTADA function

covariates

Vector of strings indicating names of covariates in propensity model

effectModifiers

Vector of strings indicating names of effect modifiers in participation model. When it is NULL, the prePostTable shows all the results

...

Further arguments from previous function or to pass to next function

x

summary.transportTADA object.

out

Output stream.

Value

The summary.transportTADA function returns a summary.transportTADA object containing the following components:

propensitySMD: Table of unweighted and weighted SMDs of covariates between treatment groups. Only propensity weights are used.

prePostTable: Table of unweighted and weighted aggregate level summary as well as pre- and post- difference values of effect modifiers of interest between study data and target data.

msmSummary: Summary object of model object for MSM. The correct variance estimators are set here for glm, whereas they are set in transportTADA for survreg and coxph.