Transportability analysis using interpolated g-computation

Estimates the average treatment effect (ATE) using interpolated g-computation in a generalizability or transportability analysis. In particular, the estimators should be unbiased for the ATE in the superpopulation or the target population, respectively.

Usage

transportInterpolated(
  link = c("identity", "log"),
  effectModifiers,
  mainTreatmentEffect,
  mainSE,
  subgroupTreatmentEffects,
  subgroupSEs,
  corrStructure = NULL,
  studySampleSize,
  aggregateStudyData,
  targetData
)

Arguments

link

Defaults to "identity", which corresponds to absolute treatment effects for continuous responses. The "log" option accommodates relative treatment effects such as relative risk, odds ratio and hazard ratio.

effectModifiers

Vector of strings indicating effect modifiers to adjust for

mainTreatmentEffect

Estimate of ATE in original study

mainSE

Estimate of standard error of estimator of ATE in original study

subgroupTreatmentEffects

Estimates of subgroup ATEs in original study. Please provide subgroup ATEs in the order of effect modifiers listed in effectModifiers, and provide the ATE of the subgroup whose proportion is provided in summaryAggregateData first in each pair

subgroupSEs

Estimates of standard errors of subgroup ATEs in original study. Please provide SEs in the order of effect modifiers listed in effectModifiers, and provide the SE of the subgroup whose proportion is provided in summaryAggregateData first in each pair

corrStructure

Correlation structure of dichotomized effect modifiers. If target IPD is provided, this will be estimated from the target data, if user input is omitted. If target aggregate data is provided, this will be specified by the user and default to an independent correlation structure if left unspecified.

studySampleSize

Sample size of original study

aggregateStudyData

Vector of proportions of dichotomized effect modifiers in study data. Please provide proportions of only one category for each effect modifier. This category should correspond to the the first ATE and SE provided for each effect modifier.

targetData

May be IPD or aggregate. If aggregate, provide proportions of only one category of dichotomized effect modifiers in a named vector (not a data frame)

Value

A transportInterpolated object with the following components:

effect: Transported ATE estimate

link: Denotes the link function used. Absolute treatment effects (i.e. for continuous outcomes) correspond to "identity", while relative treatment effects correspond to "log"

var: Transported variance estimate of effect estimate

effectModifiers: Vector of strings indicating effect modifiers adjusted for

mainTreatmentEffect: Estimate of ATE in original study

mainSE: Standard error of estimator of ATE in original study

subgroupTreatmentEffects: Estimates of subgroup ATEs in original study, as provided

subgroupSEs: Estimates of standard errors of subgroup ATEs in original study, as provided

corrStructure: Correlation structure of effect modifiers used in analysis

studySampleSize: Sample size of original study

aggregateStudyData: Aggregate-level study data, as provided

targetData: Target data, as provided

Details

This function transports the ATE estimate from the original study data to the target data by utilizing subgroup effect estimates in the same way as network meta-interpolation (Harari et al., 2023). As standard errors are transported manually, no bootstrapping is done, unlike other methods supported by TransportHealthR.

References

Harari O, Soltanifar M, Cappelleri JC, et al. Network meta-interpolation: Effect modification adjustment in network meta-analysis using subgroup analyses. Res Syn Meth. 2023; 14(2): 211-233. doi:10.1002/jrsm.1608